requirements.dev.txt
pytest==8.2.0
pytest-cov==5.0.0
pandas==2.2.2
numpy==1.26.4
faker==25.0.0
langchain-core>=0.2.20
langchain-ollama>=0.1.0
chromadb==0.5.0
python-dotenv==1.0.1